+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './aspect';
 export * from './construct';
 export * from './metadata';
+export * from './dependency';